added new make target to start nootka without install, fixed confirm tip appearing and pitch detection during exam launching

+++ packaging/src/make-src.sh
@@ -36,7 +36,12 @@ cp $SRC_DIR/gpl $DST_DIR
 cp $SRC_DIR/README $DST_DIR
 cp $SRC_DIR/TODO $DST_DIR
 cp $SRC_DIR/cmake_uninstall.cmake.in $DST_DIR
-cp $SRC_DIR/.hg/cache/branchheads $DST_DIR/.hg/cache/
+if [ -f $SRC_DIR/.hg/cache/branchheads ]; then
+		cp $SRC_DIR/.hg/cache/branchheads $DST_DIR/.hg/cache/
+fi
+if [ -f $SRC_DIR/.hg/cache/branchheads-served ]; then
+		cp $SRC_DIR/.hg/cache/branchheads-served $DST_DIR/.hg/cache/branchheads
+fi
 
 tar -cjf $DST_DIR.tar.bz2 $DST_DIR
